Добавлена функциональность "Отзывы". Тесты Постман
