+++ src/main/java/ru/yandex/practicum/filmorate/mapper/ReviewMapper.java
@@ -5,10 +5,17 @@ import ru.yandex.practicum.filmorate.dto.ReviewDto;
 import ru.yandex.practicum.filmorate.model.Review;
 
 /**
- * Маппер для преобразования между Review и ReviewDto.
+ * Маппер для преобразования между DTO и моделью отзыва.
+ * Обеспечивает конвертацию объектов ReviewDto в Review и обратно.
  */
 @Component
 public class ReviewMapper {
+    /**
+     * Преобразует ReviewDto в Review.
+     *
+     * @param reviewDto DTO отзыва
+     * @return объект модели Review
+     */
     public Review toReview(ReviewDto reviewDto) {
         Review review = new Review();
         review.setContent(reviewDto.getContent());
@@ -18,6 +25,12 @@ public class ReviewMapper {
         return review;
     }
 
+    /**
+     * Преобразует Review в ReviewDto.
+     *
+     * @param review объект модели Review
+     * @return DTO отзыва
+     */
     public ReviewDto toReviewDto(Review review) {
         ReviewDto reviewDto = new ReviewDto();
         reviewDto.setContent(review.getContent());
